Added HttpContextAccessor service and session middleware
The `HttpContextAccessor` service has been added to the application's services collection to allow accessing the `HttpContext` outside of the controller context. A new session middleware with a 30-minute idle timeout, a 30-minute maximum cookie age, a cookie name of ".DiabetesApp.Session", and a secure policy has been added. The `UseSession` middleware has been added to the application's middleware pipeline to enable session state, which must be added before `UseRouting` and `UseEndpoints`.